first_try

+++ DeepK/Data.Class.cpp
@@ -1,5 +1,5 @@
 #include "Data.Class.hpp"
 
 DataClass::DataClass( void ) {
-	_vF = new std::vector<uint64_t>(static_cast<uint64_t>(std::numeric_limits<int>::max()));
+	_vF = new std::vector<uint64_t>(static_cast<uint64_t>(std::numeric_limits<uint64_t>::max()));
 }
